fix: rename cmd/halhealth → cmd/health, package halhealth → health
